% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{calcCost}
\alias{calcCost}
\title{Calculates the cost of the design.
Supports costPowerOptimise which is in Beta and not currently implemented}
\usage{
calcCost(x)
}
\description{
Calculates the cost of the design.
Supports costPowerOptimise which is in Beta and not currently implemented
}
\examples{
dataComponents<-do.call(designFactors,list(toolbox.interface.file=excelInFile),quote=TRUE)
initialCost<-calcCost(scenario.matrix,TRUE)
calcCost(as.numeric(startScenario[c(1,2,5,7)]))  # should equal 0
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (in review) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
