% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{costPowerOptimise}
\alias{costPowerOptimise}
\title{Optimise Power for Cost}
\usage{
costPowerOptimise(scenarioIndex = 1)
}
\description{
Currently in Beta and not implemented
}
\examples{
excelInFile<-"epower_interface_V1.3.xlsx"
ncores<-1
dir.name=getwd()
require(doParallel,quietly=TRUE)
require(foreach,quietly=TRUE)
dataComponents<-do.call(designFactors,list(toolbox.interface.file=excelInFile),quote=TRUE)
scenarioParams<-do.call(powerScenario,list(inputData=dataComponents))
scenario.matrix<-do.call(buildScenarioMatrix,list(params=scenarioParams))
scenarioEvaluation<-costPowerOptimise(startScenario)
}
