% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{designFactors}
\alias{designFactors}
\title{Data Object Extraction}
\usage{
designFactors(toolbox.interface.file = excelInFile, read_type = read_package)
}
\arguments{
\item{toolbox.interface.file}{the name of the excel file}
}
\value{
A list of design specification parameters
}
\description{
Read excel scenario workbook and convert to R data objects
}
\details{
This function uses XLConnect or openxlsx to read the excel scenario
workbook, and then processes that information into
a series of R objects.
}
\examples{
dataComponents<-designFactors()
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (2019) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution.
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
