% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\docType{package}
\name{epower-package}
\alias{epower-package}
\alias{epower}
\title{\tabular{ll}{
Package: \tab epower\cr
Type: \tab Package\cr
Title: \tab Power analysis for beyond BACI sampling designs\cr
Version: \tab 1.3\cr
Date: \tab 2018-10-30\cr
Author: \tab BMT, AIMS and QUT (Rebecca Fisher, Glenn Shiell, Rohan Sadler, James McGree)\cr
Maintainer: \tab Rebecca Fisher\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
Depends: \tab doParallel, INLA\cr
Suggests: \tab openxlsx, XLConnect\Cr
}}
\details{
epower conducts a power analysis on beyond BACI sampling
designs through INLA and Monte Carlo simulation.

Currently epower considers only univariate responses. The analysis allows for
crossed random effects between \code{Time} and \code{Site} in the design.

The epower toolbox has been designed for use by ecologists and environmental practioners
with a sound understanding of the statistical principles of
BACI designs. To assist the non-R user, BACI parameters and pilot data are
defined in an Excel workbook. The template epower_interface_V1.3.xlsx should
have been supplied with the package tar.gz file.

To run the program, simply adjust your BACI design
in a copy of epower_interface_V1.3.xlsx, load the
package and then call the function fitData(), which extracts the relevant information
from the excel interface, followed by assessPower(), which carries out the power analysis
based on the scenario information extracted.
Alternatively, the program can be run entirely within R, using the function supplyData
and calling power Scenario manually before running assessPower().
The epower (V1.3) package is a BMT product which has been developed in
collaboration with the Australian Institute of Marine Science (AIMS),
Queensland University of Technology (QUT) and Pink Lake Analytics.
The method/code supporting epower v1.3 is described in Fisher R et al. (2019, Methods in Ecology and Evolution).
Detailed instructions for operating the package can be found in the user manual (BMT 2019), available as Appendix 1 (Fisher et al 2019), or from BMTglobal.
epower is freely available to users under licence.
The licence explicitly precludes the incorporation of the epower code into
other programs which are subsequently sold or used for commercial purposes.
Users are asked to please include the following citation text:
Analyses were carried out using the software epower V1.3 (BMT 2019)
as described in Fisher et al (2019) and (BMT 2019), based on the
statistical programming platform R (R-Core Team, 2019).
Acknowledgements:

BMT funded the development
of this package with Glenn Shiells as project leader. Rebecca
Fisher (AIMS, Perth, Western Australia) encoded the toolbox and
developed the formulation of the BACI design. James McGree (QUT, Brisbane,
Queensland) provided advice on implementation of the toolbox using INLA,
the use of Bayes factors for statistical hypothesis testing (model selection)
and the use of a posterior sample as the basis for monte-carlo simulation.
Rohan Sadler (Pink Lake Analytics, Perth, Western Australia) restructured
the code as the epower package, added parallelization and
assisted in the design of a beta cost optimisation solution (not currently implemented).

Discussion with Marti Anderson (affiliation), Carl Schwarz (Simon Fraser
University, British Columbia, Canada), Tim Langlois (University of Western
Australia, WA, Australia) and Kim Freidman(previously at the Department of Parks
and Wildlife, WA, Australia) during development of epower was highly valued.
}
\examples{
install.packages("epower",dependencies=TRUE)
library(epower)
# Set the working directory in R to the folder containg the
# excel workbook. This can be done by clicking
# File -> Change dir...

}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM
(in review) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution

BMT (2018). epower: Statistical Power Analysis for BACI designs; Guide to Software and Statistical Methods.
}
\seealso{
\code{\link{fitData}}

\code{\link{assessPower}}

\code{\link{powerScenario}}

\code{\link{supplyData}}
}
\author{
Rebecca Fisher (Australian Institue of Marine Science)

Glenn Shiell (BMT)

Rohan Sadler (Pink Lake Analytics)

James McGree (Queensland University of Technology)

Maintainer: Rebecca Fisher \email{r.fisher@aims.gov.au}
}
\keyword{package}
