% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{evaluatePower}
\alias{evaluatePower}
\title{Evaluate Power of Design.
 Supports costPowerOptimise which is in Beta and not currently implemented}
\usage{
evaluatePower(x, ncores = dataComponents$ncores)
}
\description{
Evaluate Power of Design.
 Supports costPowerOptimise which is in Beta and not currently implemented
}
\examples{
excelInFile<-"epower_interface_V1.3.xlsx"
ncores<-1
dir.name=getwd()
require(INLA)
require(doParallel,quietly=TRUE)
require(foreach,quietly=TRUE)
dataComponents<-do.call(designFactors,list(toolbox.interface.file=excelInFile),quote=TRUE)
scenarioParams<-do.call(powerScenario,list(inputData=dataComponents))
scenario.matrix<-do.call(buildScenarioMatrix,list(params=scenarioParams))
startScenario<-scenario.matrix[1,]
scenarioEvaluation<-evaluatePower(as.numeric(startScenario[c(1:2,5,7)]))#,missingPars=startScenario)
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (2019) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution.
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
