% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{powerScenario}
\alias{powerScenario}
\title{Scenario Parameters}
\usage{
powerScenario(inputData = dataComponents)
}
\arguments{
\item{inputData}{the data objects extracted from the excel
file defining the power analysis scenario; by default it
inherits dataComponents from \code{\link{designFactors}}.}
}
\value{
A list with scenario defining parameters and objects, including
a posterior sample, along with separate components
for the hyper parameters (hyperpar.sample, note this includes samples for
each BACI strata) and fixed parameter estimates (fixed.sample).
The fitted inla model is also returned, along with the
random structure of the design specified.
}
\description{
Defines the power analysis scenario parameters and objects
}
\details{
Parameters controlling the power analysis to be undertaken
are returned as an object, along with model specificiations.

Inla is used to fit models, and a posterior sample
(of size n.its) is obtained that includes posterior samples for each strata of
the BACI design, each level of the fixed structure available in the pilot data
(Before v After * Control v Impact), and any other hyperparameters
specific to the statistical distribution used (e.g. random error estimates
for a Gaussian fit, estimate of theta for a negative binomial).
Also defined are variances at each strata of the BACI design
for Monte Carlo simulation if required and desired effect sizes.
}
\examples{
scenarioParams<-powerScenario()
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (2019) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution.
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
