% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{run.scenario}
\alias{run.scenario}
\title{Simulate Test}
\usage{
run.scenario(x, scenario = scenarioParams)
}
\arguments{
\item{x}{A matrix of scenario specificiations to be run}

\item{scenario}{The scenario parameters as specified in scenarioParams (generated
by a call to fitData}
}
\value{
a list of realised model probabilities for each scenario, with each
scenario having as many as required by the number of iterations. Also returns
a list of the simulated datasets used in testing.
}
\description{
Performs a Monte Carlo simulation of the BACI design based on the
posterior sample obtained from the fit of the pilot data and tests for
significant effects (via model probabilities).
}
\details{
This runs statistical testing across the posterior sample (number of requested
iterations). An approximate type I or type II (i.e. 1-power) error rate
is calculated based on the proportion of tests failing the criteria
set for model probabilities (>0.5).

Note that Type I error is simply the proportion of tests identified as
significantly different when the effect size is zero.

This power calculation includes uncertainty of parameter estimates
via use of the posterior sample obtained through INLA.
}
\examples{
#scenarioRun<-run.scenario(scenario.matrix[2,], scenarioParams)
scenarioRun<-run.scenario(x1)
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (2019) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution.
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
