% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{supplyData}
\alias{supplyData}
\title{supplyData}
\usage{
supplyData(
  dat,
  variableType = "gaussian",
  design.matrix,
  levels.dat,
  scenario.data,
  effect.info = list(Multiplicative = 1, Fixed.change = 0, Effect.values = "0;-0.3"),
  ncores = 1
)
}
\arguments{
\item{dat}{A data.frame containing the pilot data to use in the analysis}

\item{variableType}{A character string indicating the distribution to use
for the respon variable. One "gaussian", "gamma", "poisson", "nbinomial", "binomial" or "beta"}

\item{design.matrix}{A named list specifying the column headings to use from data that
correspond to different elements of the BACI design. Must contain:
Response (the column (heading) name in dat corresponding to the variable for which power is the be examined),
Trials (For a binomial response the column in dat indicating the number of trials),
Location, sublocation , Time, subtime,
BvA (the column (factor) in dat indicating data from "before" versus "after" the impact),
CvI (the column (factor) in dat indicating data from "control" (or reference) versus "impact" sites).
Elements not relevant to the design
must be included in the list, but may be specified as NA (e.g. subtime, sublocation).}

\item{levels.dat}{A named list containing the elements:
Before (the factor level in the BvA column of dat used to indicate the before data),
Control (the factor level in the CvI column of dat used to indicate data
for the control or reference location(s)),
After (the factor level in the BvA column of dat used to indicate the after data), and
Impact (the factor level in the CvI column of dat used to indicate data
for the impact location(s)).}

\item{scenario.data}{A named list containing the elements:
Number.of.iterations (the number of iterations to perform.
Note: It is recommended to run a small number (~5) initially
to confirm the simulation is working. Preliminary results
can be examined using ~100 iterations,
but a minimum of ~500 should be performed for final results),
filename="test" (a name for the simulation data to be saved as),
Number.of.Impact.Locations (the number of Impact Locations in the design),
Number.of.Control.Locations (the number of Control Locations in the design),
Number.of.sublocations.within.Location (the number of sublocations in the design.
Sublocation is nested within Location),
Number.of.sample.times.Before (the number of Before Times in the design),
Number.of.sample.times.After (the number of After Times in the design),
Number.of.subtimes.within.Time (the number of subtimes () in the design.
Subtime () is nested within Time, if there are no subtimes specify as 1),
Number.of.trials (the number of trials where data are a proportion or frequency,
and a binomial distribution is used),
Number.of.replicate.measurements (the number of replicates to use.
Note that a fixed replicate design is specified with replication = 1,
and the replicate ID assigned at "sublocation" above).
Separate multiple values to test using ";". E.g. 3;6;12.}

\item{effect.info}{A names list containing the elements:
Multiplicative (1, if a multiplicative effect is desired, otherwise 0),
Fixed.change (1, if a Fixed (additive) effect is desired, otherwise 0),
Effect.values=-0.3 (For a multiplicative effect: 0 = no effect; -0.1= a 10% decline;
+0.1 = a 10% increase etc.
For an absolute change in the fixed effect: 0 = no effect; +10 = an increase of 10 units;
and -10 = a decrease of 10 units).
Separate multiple effect size values to test using ";".
Note: "Multiplicative"
will multiply the response variable by the supplied value and then add this as the
"effect"; "Fixed" will add the supplied value to the response variable directly.
For either case the "impact" is applied only to "after" "impact" samples.}

\item{ncores}{the number of cores required for the analysis

Unlike fitData(), the function supplyData() only generates the dataComponents
object and does not automatically call powerScenario, this step must be done
manually by the user.

You will need to check the current approximate CPU load
on the available cores before implementing ncores > 1}
}
\value{
A dataComponents list containing all the information required to run powerScenario.
}
\description{
An alternative to fitData() that allows the power analysis to be set up and run
entirely through R, without using the excel interface.
It generates the necessary dataComponents object required to run powerScenario,
which generates the necessary object scenarioParams to run function assessPower.
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (2019) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution.
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
