% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolbox_run.r
\name{xformEta}
\alias{xformEta}
\title{Transform ETA}
\usage{
xformEta(x, variableType, direction)
}
\arguments{
\item{x}{- a number or numeric vector on which to apply the transformation}

\item{variableType}{- the statsitical distirbution being used}

\item{direction}{- whether to go from the "link" to the "response" scale}
}
\value{
A number or numeric vector of the transformed result
}
\description{
A function to transform between the response and link scales as required
}
\examples{
xformEta(c(0.2,0.5,0.8),"binomial","link")
xformEta(c(-1,0,1),"binomial","response")
}
\references{
Fisher R, Shiell GR, Sadler RJ, Inostroza K, Shedrawi G, Holmes TH, McGree JM (in review) epower: an R package for power analysis of Before-After-Control-Impact (BACI) designs. Methods in Ecology and Evolution
}
\author{
Rebecca Fisher \email{r.fisher@aims.gov.au}
}
